MatchIt: Nonparametric Preprocessing for Parametric Causal Inference

Selects matched samples of the original treated and control groups with similar covariate distributions – can be used to match exactly on covariates, to match on propensity scores, or perform a variety of other matching procedures. The package also implements a series of recommendations offered in Ho, Imai, King, and Stuart (2007) doi:10.1093/pan/mpl013 . (The 'gurobi' package, which is not on CRAN, is optional and comes with an installation of the Gurobi Optimizer, available at https://www.gurobi.com.)

See also

Author

Maintainer: Noah Greifer noah.greifer@gmail.com (ORCID)

Authors:

Daniel Ho daniel.e.ho@gmail.com (ORCID)

Kosuke Imai imai@harvard.edu (ORCID)

Gary King king@harvard.edu (ORCID)

Elizabeth Stuart estuart@jhu.edu (ORCID)

Other contributors:

Alex Whitworth whitworth.alex@gmail.com [contributor]