Cardinality Matching

In matchit(), setting method = "cardinality" performs cardinality matching and other forms of matching using mixed integer programming. Rather than forming pairs, cardinality matching selects the largest subset of units that satisfies user-supplied balance constraints on mean differences. One of several available optimization programs can be used to solve the mixed integer program. The default is the GLPK library as implemented in the Rglpk package, but performance can be dramatically improved using Gurobi and the gurobi package, for which there is a free academic license.

This page details the allowable arguments with method = "cardinality". See matchit() for an explanation of what each argument means in a general context and how it can be specified.

Below is how matchit() is used for cardinality matching:

matchit(formula,
        data = NULL,
        method = "cardinality",
        estimand = "ATT",
        exact = NULL,
        discard = "none",
        s.weights = NULL,
        ratio = 1,
        verbose = FALSE, ...)

Arguments

formula
a two-sided formula() object containing the treatment and covariates to be balanced.

data
a data frame containing the variables named in formula. If not found in data, the variables will be sought in the environment.

method
set here to "cardinality".

estimand
a string containing the desired estimand. Allowable options include "ATT", "ATC", and "ATE". See Details.

exact
for which variables exact matching should take place. Separate optimization will occur within each subgroup of the exact matching variables.

discard
a string containing a method for discarding units outside a region of common support.

s.weights
the variable containing sampling weights to be incorporated into the optimization. The balance constraints refer to the product of the sampling weights and the matching weights, and the sum of the product of the sampling and matching weights will be maximized.

ratio
the desired ratio of control to treated units. Can be set to NA to maximize sample size without concern for this ratio. See Details.

verbose
logical; whether information about the matching process should be printed to the console.

...
additional arguments that control the matching specification:

tols

numeric; a vector of imbalance tolerances for mean differences, one for each covariate in formula. If only one value is supplied, it is applied to all. See std.tols below. Default is .05 for standardized mean differences of at most .05 for all covariates between the treatment groups in the matched sample.

std.tols

logical; whether each entry in tols corresponds to a raw or standardized mean difference. If only one value is supplied, it is applied to all. Default is TRUE for standardized mean differences. The standardization factor is the pooled standard deviation when estimand = "ATE", the standard deviation of the treated group when estimand = "ATT", and the standard deviation of the control group when estimand = "ATC" (the same as used in summary.matchit()).

solver

the name of solver to use to solve the optimziation problem. Available options include "glpk", "symphony", and "gurobi" for GLPK (implemented in the Rglpk package), SYMPHONY (implemented in the Rsymphony package), and Gurobi (implemented in the gurobi package). The differences between them are in speed and solving ability. GLPK (the default) is the easiest to install, but Gurobi is recommended as it consistently outperforms other solvers and can find solutions even when others can't, and in less time. Gurobi is proprietary but can be used with a free trial or academic license. SYMPHONY may not produce reproducible results, even with a seed set.

time

the maximum amount of time before the optimization routine aborts, in seconds. Default is 120 (2 minutes). For large problems, this should be set much higher.

Outputs

Most outputs described in matchit() are returned with method = "cardinality". The match.matrix and subclass components are omitted because no pairing or subclassification is done. When include.obj = TRUE in the call to matchit(), the output of the optimization function will be included in the output. When exact is specified, this will be a list of such objects, one for each stratum of the exact variables.

Details

Cardinality and Template Matching

Two types of matching are available with method = "cardinality": cardinality matching and template matching.

Cardinality matching finds the largest matched set that satisfies the balance constraints between treatment groups, with the additional constraint that the ratio of the number of matched control to matched treated units is equal to ratio (1 by default), mimicking k:1 matching. When not all treated units are included in the matched set, the estimand no longer corresponds to the ATT, so cardinality matching should be avoided if retaining the ATT is desired. To request cardinality matching, estimand should be set to "ATT" or "ATC" and ratio should be set to a positive integer. 1:1 cardinality matching is the default method when no arguments are specified.

Template matching finds the largest matched set that satisfies balance constraints between each treatment group and a specified target sample. When estimand = "ATT", it will find the largest subset of the control units that satisfies the balance constraints with respect to the treated group, which is left intact. When estimand = "ATE", it will find the largest subsets of the treated group and of the control group that are balanced to the overall sample. To request template matching for the ATT, estimand should be set to "ATT" and "ratio" to NA. To request template matching for the ATE, estimand should be set to "ATE" and ratio can be set either to NA to maximize the size of each sample independently or to a positive integer to ensure that the ratio of matched control units to matched treated treats is fixed, mimicking k:1 matching. Unlike cardinality matching, template matching retains the requested estimand if a solution is found.

Neither method involves creating pairs in the matched set, but it is possible to perform an additional round of pairing within the matched sample after cardinality matching or template matching for the ATE with a fixed sample size ratio. See Examples for an example of optimal pair matching after cardinality matching. The balance will not change, but additional precision and robustness can be gained by forming the pairs.

The weights are scaled so that the sum of the weights in each group is equal to the number of matched units in the smaller group when cardinality matching or template matching for the ATE, and scaled so that the sum of the weights in the control group is equal to the number of treated units when template matching for the ATT. When the sample sizes of the matched groups is the same (i.e., when ratio = 1), no scaling is done. Robust standard errors should be used in effect estimation after cardinality or template matching (and cluster-robust standard errors if additional pairing is done in the matched sample). See vignette("estimating-effects") for more information.

Specifying Balance Constraints

The balance constraints are on the (standardized) mean differences between the matched treatment groups for each covariate. Balance constraints should be set by supplying arguments to tols and std.tols. For example, setting tols = .1 and std.tols = TRUE requests that all the mean differences in the matched sample should be within .1 standard deviations for each covariate. Different tolerances can be set for difference variables; it might be beneficial to constrain the mean differences for highly prognostic covariates more tightly than other variables. For example, one could specify tols = c(.001, .05), std.tols = c(TRUE, FALSE) to request that the standadized mean difference for the first covariate is less than .001 and the raw mean difference for the second covariate is less than .05. The values should be specified in the order they appear in formula, except when interactions are present. One can run the following code:

MatchIt:::get_assign(model.matrix(~X1*X2 + X3,
                                  data = data))[-1]

which will output a vector of numbers and the variable to which each number corresponds; the first entry in tols corresponds to the variable labeled 1, the second to the variable labeled 2, etc.

Dealing with Errors and Warnings

When the optimization cannot be solved at all, or at least within the time frame specified in the argument to time, an error or warning will appear. Unfortunately, it is hard to know exactly the cause of the failure and what measures should be taken to rectify it.

A warning that says "The optimizer failed to find an optimal solution in the time alotted. The returned solution may not be optimal." usually means that an optimal solution may be possible to find with more time, in which case time should be increased or a faster solver should be used. Even with this warning, a potentially usable solution will be returned, so don't automatically take it to mean the optimization failed. Sometimes, when there are multiple solutions with the same resulting sample size, the optimizers will stall at one of them, not thinking it has found the optimum. The result should be checked to see if it can be used as the solution.

An error that says "The optimization problem may be infeasible." usually means that there is a issue with the optimization problem, i.e., that there is no possible way to satisfy the constraints. To rectify this, one can try relaxing the constraints by increasing the value of tols or use another solver. Sometimes Gurobi can solve problems that the other solvers cannot.

References

In a manuscript, you should reference the solver used in the optimization. For example, a sentence might read:

Cardinality matching was performed using the MatchIt package (Ho, Imai, King, & Stuart, 2011) in R with the optimization performed by GLPK.

See vignette("matching-methods") for more literature on cardinality matching.

See also

Examples

data("lalonde")

#Choose your solver; "gurobi" is best, "glpk" is free and
#easiest to install
solver <- "glpk"

# 1:1 cardinality matching
m.out1 <- matchit(treat ~ age + educ + re74 + re75,
                  data = lalonde, method = "cardinality",
                  estimand = "ATT", ratio = 1,
                  tols = .1, solver = solver)
m.out1

#> A matchit object
#>  - method: Cardinality matching
#>  - number of obs.: 614 (original), 370 (matched)
#>  - target estimand: ATT
#>  - covariates: age, educ, re74, re75

summary(m.out1)

#> 
#> Call:
#> matchit(formula = treat ~ age + educ + re74 + re75, data = lalonde, 
#>     method = "cardinality", estimand = "ATT", ratio = 1, tols = 0.1, 
#>     solver = solver)
#> 
#> Summary of Balance for All Data:
#>      Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean eCDF Max
#> age        25.8162       28.0303         -0.3094     0.4400    0.0813   0.1577
#> educ       10.3459       10.2354          0.0550     0.4959    0.0347   0.1114
#> re74     2095.5737     5619.2365         -0.7211     0.5181    0.2248   0.4470
#> re75     1532.0553     2466.4844         -0.2903     0.9563    0.1342   0.2876
#> 
#> 
#> Summary of Balance for Matched Data:
#>      Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean eCDF Max
#> age        25.8162       25.8649         -0.0068     0.4943    0.0688   0.2054
#> educ       10.3459       10.2000          0.0726     0.3979    0.0447   0.1135
#> re74     2095.5737     2570.5264         -0.0972     0.9780    0.0615   0.3243
#> re75     1532.0553     1852.2194         -0.0995     1.1566    0.0594   0.2054
#> 
#> Percent Balance Improvement:
#>      Std. Mean Diff. Var. Ratio eCDF Mean eCDF Max
#> age             97.8       14.2      15.4    -30.2
#> educ           -32.1      -31.4     -28.6     -1.9
#> re74            86.5       96.6      72.6     27.5
#> re75            65.7     -225.6      55.7     28.6
#> 
#> Sample Sizes:
#>           Control Treated
#> All           429     185
#> Matched       185     185
#> Unmatched     244       0
#> Discarded       0       0
#> 

# Template matching for the ATT
m.out2 <- matchit(treat ~ age + educ + re74 + re75,
                  data = lalonde, method = "cardinality",
                  estimand = "ATT", ratio = NA,
                  tols = .1, solver = solver)
m.out2

#> A matchit object
#>  - method: Cardinality matching
#>  - number of obs.: 614 (original), 521 (matched)
#>  - target estimand: ATT
#>  - covariates: age, educ, re74, re75

summary(m.out2, un = FALSE)

#> 
#> Call:
#> matchit(formula = treat ~ age + educ + re74 + re75, data = lalonde, 
#>     method = "cardinality", estimand = "ATT", ratio = NA, tols = 0.1, 
#>     solver = solver)
#> 
#> Summary of Balance for Matched Data:
#>      Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean eCDF Max
#> age        25.8162       25.9643         -0.0207     0.4880    0.0687   0.1794
#> educ       10.3459       10.1518          0.0966     0.4933    0.0348   0.1046
#> re74     2095.5737     2583.5615         -0.0999     2.2916    0.1137   0.3748
#> re75     1532.0553     1739.6074         -0.0645     1.6897    0.0683   0.2310
#> 
#> Sample Sizes:
#>           Control Treated
#> All           429     185
#> Matched       336     185
#> Unmatched      93       0
#> Discarded       0       0
#> 

# Template matching for the ATE with a constraint on
# sample size
m.out3 <- matchit(treat ~ age + educ + re74 + re75,
                  data = lalonde, method = "cardinality",
                  estimand = "ATE", ratio = 1,
                  tols = .1, solver = solver)
m.out3

#> A matchit object
#>  - method: Cardinality matching
#>  - number of obs.: 614 (original), 180 (matched)
#>  - target estimand: ATE
#>  - covariates: age, educ, re74, re75

summary(m.out3, un = FALSE)

#> 
#> Call:
#> matchit(formula = treat ~ age + educ + re74 + re75, data = lalonde, 
#>     method = "cardinality", estimand = "ATE", ratio = 1, tols = 0.1, 
#>     solver = solver)
#> 
#> Summary of Balance for Matched Data:
#>      Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean eCDF Max
#> age        27.2778       27.7000         -0.0461     0.5835    0.0506   0.1000
#> educ       10.1667       10.1667          0.0000     0.6634    0.0503   0.1444
#> re74     4276.6254     4815.2682         -0.0911     0.8934    0.0306   0.1111
#> re75     2346.7966     2345.2431          0.0005     1.0377    0.0470   0.1111
#> 
#> Sample Sizes:
#>           Control Treated
#> All           429     185
#> Matched        90      90
#> Unmatched     339      95
#> Discarded       0       0
#> 

# Pairing after 1:1 cardinality matching:
m.out4 <- matchit(treat ~ age + educ + re74 + re75,
                  data = lalonde, method = "nearest",
                  distance = "mahalanobis",
                  discard = m.out1$weights == 0)

# Note that balance doesn't change but pair distances
# are lower for the paired-upon variables
summary(m.out4, un = FALSE)

#> 
#> Call:
#> matchit(formula = treat ~ age + educ + re74 + re75, data = lalonde, 
#>     method = "nearest", distance = "mahalanobis", discard = m.out1$weights == 
#>         0)
#> 
#> Summary of Balance for Matched Data:
#>      Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean eCDF Max
#> age        25.8162       25.8649         -0.0068     0.4943    0.0688   0.2054
#> educ       10.3459       10.2000          0.0726     0.3979    0.0447   0.1135
#> re74     2095.5737     2570.5264         -0.0972     0.9780    0.0615   0.3243
#> re75     1532.0553     1852.2194         -0.0995     1.1566    0.0594   0.2054
#>      Std. Pair Dist.
#> age           0.6520
#> educ          0.6479
#> re74          0.4648
#> re75          0.5263
#> 
#> Sample Sizes:
#>           Control Treated
#> All           429     185
#> Matched       185     185
#> Unmatched       0       0
#> Discarded     244       0
#> 

summary(m.out1, un = FALSE)

#> 
#> Call:
#> matchit(formula = treat ~ age + educ + re74 + re75, data = lalonde, 
#>     method = "cardinality", estimand = "ATT", ratio = 1, tols = 0.1, 
#>     solver = solver)
#> 
#> Summary of Balance for Matched Data:
#>      Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean eCDF Max
#> age        25.8162       25.8649         -0.0068     0.4943    0.0688   0.2054
#> educ       10.3459       10.2000          0.0726     0.3979    0.0447   0.1135
#> re74     2095.5737     2570.5264         -0.0972     0.9780    0.0615   0.3243
#> re75     1532.0553     1852.2194         -0.0995     1.1566    0.0594   0.2054
#> 
#> Sample Sizes:
#>           Control Treated
#> All           429     185
#> Matched       185     185
#> Unmatched     244       0
#> Discarded       0       0
#> 

# In these examples, a high tol was used and
# few covariate matched on in order to not take too long;
# with real data, tols should be much lower and more
# covariates included if possible.