Nearest Neighbor Matching

In matchit(), setting method = "nearest" performs greedy nearest neighbor matching. A distance is computed between each treated unit and each control unit, and, one by one, each treated unit is assigned a control unit as a match. The matching is "greedy" in the sense that there is no action taken to optimize an overall criterion; each match is selected without considering the other matches that may occur subsequently.

This page details the allowable arguments with method = "nearest". See matchit() for an explanation of what each argument means in a general context and how it can be specified.

Below is how matchit() is used for nearest neighbor matching:

matchit(formula,
        data = NULL,
        method = "nearest",
        distance = "glm",
        link = "logit",
        distance.options = list(),
        estimand = "ATT",
        exact = NULL,
        mahvars = NULL,
        antiexact = NULL,
        discard = "none",
        reestimate = FALSE,
        s.weights = NULL,
        replace = TRUE,
        m.order = NULL,
        caliper = NULL,
        ratio = 1,
        min.controls = NULL,
        max.controls = NULL,
        verbose = FALSE, ...)

Arguments

formula
a two-sided formula() object containing the treatment and covariates to be used in creating the distance measure used in the matching.

data
a data frame containing the variables named in formula. If not found in data, the variables will be sought in the environment.

method
set here to "nearest".

distance
the distance measure to be used. See distance for allowable options. Can be supplied as a distance matrix.

link
when distance is specified as a string and not "mahalanobis", an additional argument controlling the link function used in estimating the distance measure. See distance for allowable options with each option.

distance.options
a named list containing additional arguments supplied to the function that estimates the distance measure as determined by the argument to distance.

estimand
a string containing the desired estimand. Allowable options include "ATT" and "ATC". See Details.

exact
for which variables exact matching should take place.

mahvars
for which variables Mahalanobis distance matching should take place when a distance measure other than "mahalanobis" is used (e.g., for caliper matching or to discard units for common support). If specified, the distance measure will not be used in matching.

antiexact
for which variables ant-exact matching should take place.

discard
a string containing a method for discarding units outside a region of common support. Only allowed when distance is not "mahalanobis" and not a matrix.

reestimate
if discard is not "none", whether to re-estimate the propensity score in the remaining sample prior to matching.

s.weights
the variable containing sampling weights to be incorporated into propensity score models and balance statistics.

replace
whether matching should be done with replacement.

m.order
the order that the matching takes place. The default is "data" for distance = "mahalanobis" or when distance is supplied as a matrix, and "largest" otherwise. See matchit() for allowable options.

caliper
the width(s) of the caliper(s) used for caliper matching. See Details and Examples.

std.caliper
logical; when calipers are specified, whether they are in standard deviation units (TRUE) or raw units (FALSE).

ratio
how many control units should be matched to each treated unit for k:1 matching. For variable ratio matching, see section "Variable Ratio Matching" in Details below.

min.controls, max.controls
for variable ratio matching, the minimum and maximum number of controls units to be matched to each treated unit. See section "Variable Ratio Matching" in Details below.

verbose
logical; whether information about the matching process should be printed to the console. When TRUE, a progress bar implemented using RcppProgress will be displayed.

...
... additional arguments that control the matching specification:

reuse.max

numeric; the maximum number of times each control can be used as a match. Setting reuse.max = 1 correpsonds to matching without replacement (i.e., replace = FALSE), and setting reuse.max = Inf corresponds to traditional matching with replacement (i.e., replace = TRUE) with no limit on the number of times each control unit can be matched. Other values restrict the number of times each control can be matched when matching with replacement. replace is ignored when reuse.max is specified.