Add sampling weights to a matchit object

Adds sampling weights to a matchit object so that they are incorporated into balance assessment and creation of the weights. This would typically only be used when an argument to s.weights was not supplied to matchit() (i.e., because they were not to be included in the estimation of the propensity score) but sampling weights are required for generalizing an effect to the correct population. Without adding sampling weights to the matchit object, balance assessment tools (i.e., summary.matchit() and plot.matchit()) will not calculate balance statistics correctly, and the weights produced by match.data() and get_matches() will not incorporate the sampling weights.

add_s.weights(m, s.weights = NULL, data = NULL)

Arguments

m

a matchit object; the output of a call to matchit(), typically with the s.weights argument unspecified.

s.weights

an numeric vector of sampling weights to be added to the matchit object. Can also be specified as a string containing the name of variable in data to be used or a one-sided formula with the variable on the right-hand side (e.g., ~ SW).

data

a data frame containing the sampling weights if given as a string or formula. If unspecified, add_s.weights() will attempt to find the dataset using the environment of the matchit object.

Value

a matchit object with an s.weights component containing the supplied sampling weights. If s.weights = NULL, the original matchit object is returned.

See also

Author

Noah Greifer

Examples

data("lalonde")

# Generate random sampling weights, just
# for this example
sw <- rchisq(nrow(lalonde), 2)

# NN PS match using logistic regression PS that doesn't
# include sampling weights
m.out <- matchit(treat ~ age + educ + race + nodegree +
                   married  + re74 + re75, data = lalonde)

m.out
#> A matchit object
#>  - method: 1:1 nearest neighbor matching without replacement
#>  - distance: Propensity score
#>              - estimated with logistic regression
#>  - number of obs.: 614 (original), 370 (matched)
#>  - target estimand: ATT
#>  - covariates: age, educ, race, nodegree, married, re74, re75

# Add s.weights to the matchit object
m.out <- add_s.weights(m.out, sw)

m.out #note additional output
#> A matchit object
#>  - method: 1:1 nearest neighbor matching without replacement
#>  - distance: Propensity score
#>              - estimated with logistic regression
#>              - sampling weights not included in estimation
#>  - number of obs.: 614 (original), 370 (matched)
#>  - sampling weights: present
#>  - target estimand: ATT
#>  - covariates: age, educ, race, nodegree, married, re74, re75

# Check balance; note that sample sizes incorporate
# s.weights
summary(m.out, improvement = FALSE)
#> 
#> Call:
#> matchit(formula = treat ~ age + educ + race + nodegree + married + 
#>     re74 + re75, data = lalonde)
#> 
#> Summary of Balance for All Data:
#>            Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean
#> distance          0.5735        0.1775          1.8590     0.8843    0.3683
#> age              25.0576       27.5941         -0.4268     0.3136    0.0996
#> educ             10.6753       10.1042          0.2894     0.4768    0.0402
#> raceblack         0.8583        0.1937          1.9057          .    0.6646
#> racehispan        0.0336        0.1340         -0.5576          .    0.1005
#> racewhite         0.1081        0.6723         -1.8167          .    0.5642
#> nodegree          0.6037        0.6061         -0.0049          .    0.0024
#> married           0.1815        0.4884         -0.7962          .    0.3069
#> re74           2421.3890     5383.1595         -0.5716     0.6372    0.1922
#> re75           1660.0466     2484.6560         -0.2870     0.7909    0.1068
#>            eCDF Max
#> distance     0.6721
#> age          0.2206
#> educ         0.1452
#> raceblack    0.6646
#> racehispan   0.1005
#> racewhite    0.5642
#> nodegree     0.0024
#> married      0.3069
#> re74         0.3772
#> re75         0.2415
#> 
#> Summary of Balance for Matched Data:
#>            Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean
#> distance          0.5735        0.3465          1.0658     0.6963    0.1331
#> age              25.0576       25.0887         -0.0052     0.3081    0.1072
#> educ             10.6753       10.4958          0.0910     0.5658    0.0319
#> raceblack         0.8583        0.4402          1.1989          .    0.4181
#> racehispan        0.0336        0.1886         -0.8604          .    0.1550
#> racewhite         0.1081        0.3712         -0.8473          .    0.2631
#> nodegree          0.6037        0.6635         -0.1224          .    0.0599
#> married           0.1815        0.1970         -0.0403          .    0.0155
#> re74           2421.3890     2101.3337          0.0618     1.7659    0.0272
#> re75           1660.0466     1730.3155         -0.0245     1.1046    0.0422
#>            eCDF Max Std. Pair Dist.
#> distance     0.4679          1.0073
#> age          0.2955          1.6780
#> educ         0.0956          1.2711
#> raceblack    0.4181          1.0694
#> racehispan   0.1550          1.4101
#> racewhite    0.2631          0.8007
#> nodegree     0.0599          0.9393
#> married      0.0155          0.8415
#> re74         0.1779          0.7512
#> re75         0.1386          0.8272
#> 
#> Sample Sizes:
#>               Control Treated
#> All (ESS)      219.25    84.6
#> All            429.     185. 
#> Matched (ESS)   96.43    84.6
#> Matched        185.     185. 
#> Unmatched      244.       0. 
#> Discarded        0.       0. 
#>