Add sampling weights to a matchit object

Adds sampling weights to a matchit object so that they are incorporated into balance assessment and creation of the weights. This would typically only be used when an argument to s.weights was not supplied to matchit() (i.e., because they were not to be included in the estimation of the propensity score) but sampling weights are required for generalizing an effect to the correct population. Without adding sampling weights to the matchit object, balance assessment tools (i.e., summary.matchit() and plot.matchit()) will not calculate balance statistics correctly, and the weights produced by match.data() and get_matches() will not incorporate the sampling weights.

add_s.weights(m, s.weights = NULL, data = NULL)

Arguments

m

a matchit object; the output of a call to matchit(), typically with the s.weights argument unspecified.

s.weights

an numeric vector of sampling weights to be added to the matchit object. Can also be specified as a string containing the name of variable in data to be used or a one-sided formula with the variable on the right-hand side (e.g., ~ SW).

data

a data frame containing the sampling weights if given as a string or formula. If unspecified, add_s.weights will attempt to find the dataset using the environment of the matchit object.

Value

a matchit object with an s.weights component containing the supplied sampling weights. The nn component containing the sample sizes before and after matching will be adjusted to incorporate the sampling weights. If s.weights = NULL, the original matchit object is returned.

Author

Noah Greifer

See also

Examples

data("lalonde")

# Generate random sampling weights, just
# for this example
sw <- rchisq(nrow(lalonde), 2)

# NN PS match using logistic regression PS that doesn't
# include sampling weights
m.out <- matchit(treat ~ age + educ + race + nodegree +
                   married  + re74 + re75, data = lalonde)

m.out
#> A matchit object
#>  - method: 1:1 nearest neighbor matching without replacement
#>  - distance: Propensity score
#>              - estimated with logistic regression
#>  - number of obs.: 614 (original), 370 (matched)
#>  - target estimand: ATT
#>  - covariates: age, educ, race, nodegree, married, re74, re75

# Add s.weights to the matchit object
m.out <- add_s.weights(m.out, sw)

m.out #note additional output
#> A matchit object
#>  - method: 1:1 nearest neighbor matching without replacement
#>  - distance: Propensity score
#>              - estimated with logistic regression
#>              - sampling weights not included in estimation
#>  - number of obs.: 614 (original), 370 (matched)
#>  - sampling weights: present
#>  - target estimand: ATT
#>  - covariates: age, educ, race, nodegree, married, re74, re75

# Check balance; note that sample sizes incorporate
# s.weights
summary(m.out, improvement = FALSE)
#> 
#> Call:
#> matchit(formula = treat ~ age + educ + race + nodegree + married + 
#>     re74 + re75, data = lalonde)
#> 
#> Summary of Balance for All Data:
#>            Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean
#> distance          0.5686        0.2174          1.5625     0.8053    0.3353
#> age              26.9482       28.2471         -0.1720     0.4644    0.0753
#> educ             10.1997       10.2232         -0.0104     0.5449    0.0317
#> raceblack         0.8447        0.2649          1.6005          .    0.5797
#> racehispan        0.0393        0.1242         -0.4369          .    0.0849
#> racewhite         0.1160        0.6109         -1.5450          .    0.4948
#> nodegree          0.6876        0.5883          0.2143          .    0.0993
#> married           0.2310        0.4942         -0.6245          .    0.2632
#> re74           2262.1724     5257.0924         -0.6509     0.4885    0.1896
#> re75           2175.3088     2557.1271         -0.0971     1.3010    0.0992
#>            eCDF Max
#> distance     0.5871
#> age          0.1171
#> educ         0.0993
#> raceblack    0.5797
#> racehispan   0.0849
#> racewhite    0.4948
#> nodegree     0.0993
#> married      0.2632
#> re74         0.3830
#> re75         0.2666
#> 
#> 
#> Summary of Balance for Matched Data:
#>            Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean
#> distance          0.5686        0.4001          0.7499     0.7925    0.1076
#> age              26.9482       26.1688          0.1032     0.4655    0.0857
#> educ             10.1997       10.5499         -0.1554     0.5357    0.0337
#> raceblack         0.8447        0.5456          0.8256          .    0.2990
#> racehispan        0.0393        0.1795         -0.7220          .    0.1402
#> racewhite         0.1160        0.2748         -0.4958          .    0.1588
#> nodegree          0.6876        0.6185          0.1490          .    0.0691
#> married           0.2310        0.2244          0.0156          .    0.0066
#> re74           2262.1724     2177.7521          0.0183     1.3574    0.0391
#> re75           2175.3088     1723.3606          0.1150     1.9467    0.0608
#>            eCDF Max Std. Pair Dist.
#> distance     0.3689          0.9545
#> age          0.2507          1.3204
#> educ         0.0873          1.1128
#> raceblack    0.2990          1.0297
#> racehispan   0.1402          1.3078
#> racewhite    0.1588          0.7763
#> nodegree     0.0691          0.9913
#> married      0.0066          0.7695
#> re74         0.2089          0.8459
#> re75         0.1739          0.6045
#> 
#> Sample Sizes:
#>               Control Treated
#> All (ESS)      206.36    77.1
#> All            429.     185. 
#> Matched (ESS)   99.92    77.1
#> Matched        185.     185. 
#> Unmatched      244.       0. 
#> Discarded        0.       0. 
#>