Add sampling weights to a matchit object

Adds sampling weights to a matchit object so that they are incorporated into balance assessment and creation of the weights. This would typically only be used when an argument to s.weights was not supplied to matchit() (i.e., because they were not to be included in the estimation of the propensity score) but sampling weights are required for generalizing an effect to the correct population. Without adding sampling weights to the matchit object, balance assessment tools (i.e., summary.matchit() and plot.matchit()) will not calculate balance statistics correctly, and the weights produced by match.data() and get_matches() will not incorporate the sampling weights.

add_s.weights(m, s.weights = NULL, data = NULL)

Arguments

m

a matchit object; the output of a call to matchit(), typically with the s.weights argument unspecified.

s.weights

an numeric vector of sampling weights to be added to the matchit object. Can also be specified as a string containing the name of variable in data to be used or a one-sided formula with the variable on the right-hand side (e.g., ~ SW).

data

a data frame containing the sampling weights if given as a string or formula. If unspecified, add_s.weights will attempt to find the dataset using the environment of the matchit object.

Value

a matchit object with an s.weights component containing the supplied sampling weights. The nn component containing the sample sizes before and after matching will be adjusted to incorporate the sampling weights. If s.weights = NULL, the original matchit object is returned.

Author

Noah Greifer

See also

Examples

data("lalonde")

# Generate random sampling weights, just
# for this example
sw <- rchisq(nrow(lalonde), 2)

# NN PS match using logistic regression PS that doesn't
# include sampling weights
m.out <- matchit(treat ~ age + educ + race + nodegree +
                   married  + re74 + re75, data = lalonde)

m.out
#> A matchit object
#>  - method: 1:1 nearest neighbor matching without replacement
#>  - distance: Propensity score
#>              - estimated with logistic regression
#>  - number of obs.: 614 (original), 370 (matched)
#>  - target estimand: ATT
#>  - covariates: age, educ, race, nodegree, married, re74, re75

# Add s.weights to the matchit object
m.out <- add_s.weights(m.out, sw)

m.out #note additional output
#> A matchit object
#>  - method: 1:1 nearest neighbor matching without replacement
#>  - distance: Propensity score
#>              - estimated with logistic regression
#>              - sampling weights not included in estimation
#>  - number of obs.: 614 (original), 370 (matched)
#>  - sampling weights: present
#>  - target estimand: ATT
#>  - covariates: age, educ, race, nodegree, married, re74, re75

# Check balance; note that sample sizes incorporate
# s.weights
summary(m.out, improvement = FALSE)
#> 
#> Call:
#> matchit(formula = treat ~ age + educ + race + nodegree + married + 
#>     re74 + re75, data = lalonde)
#> 
#> Summary of Balance for All Data:
#>            Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean
#> distance          0.6000        0.1995          2.0377     0.6527    0.3655
#> age              26.1179       27.5017         -0.1847     0.4799    0.0747
#> educ             10.2312       10.2757         -0.0203     0.6070    0.0254
#> raceblack         0.8908        0.2286          2.1229          .    0.6622
#> racehispan        0.0390        0.1380         -0.5122          .    0.0991
#> racewhite         0.0703        0.6334         -2.2029          .    0.5631
#> nodegree          0.7048        0.6186          0.1888          .    0.0861
#> married           0.1915        0.4942         -0.7692          .    0.3027
#> re74           1993.2926     5336.0318         -0.7696     0.3853    0.1991
#> re75           1431.2578     2262.9864         -0.2255     1.2560    0.1386
#>            eCDF Max
#> distance     0.6752
#> age          0.1270
#> educ         0.0861
#> raceblack    0.6622
#> racehispan   0.0991
#> racewhite    0.5631
#> nodegree     0.0861
#> married      0.3027
#> re74         0.4400
#> re75         0.2899
#> 
#> 
#> Summary of Balance for Matched Data:
#>            Means Treated Means Control Std. Mean Diff. Var. Ratio eCDF Mean
#> distance          0.6000        0.3760          1.1396     0.5638    0.1344
#> age              26.1179       24.1815          0.2585     0.5572    0.0850
#> educ             10.2312       10.5476         -0.1441     0.7084    0.0232
#> raceblack         0.8908        0.4914          1.2802          .    0.3993
#> racehispan        0.0390        0.2112         -0.8901          .    0.1722
#> racewhite         0.0703        0.2974         -0.8885          .    0.2271
#> nodegree          0.7048        0.6782          0.0582          .    0.0266
#> married           0.1915        0.2062         -0.0373          .    0.0147
#> re74           1993.2926     2149.3678         -0.0359     1.1550    0.0424
#> re75           1431.2578     1363.8536          0.0183     2.2453    0.0540
#>            eCDF Max Std. Pair Dist.
#> distance     0.4474          1.0916
#> age          0.2992          1.3311
#> educ         0.0622          1.1424
#> raceblack    0.3993          1.1957
#> racehispan   0.1722          1.3131
#> racewhite    0.2271          0.9727
#> nodegree     0.0266          1.0073
#> married      0.0147          0.8242
#> re74         0.3118          0.8961
#> re75         0.2275          0.6443
#> 
#> Sample Sizes:
#>               Control Treated
#> All (ESS)      211.46   93.93
#> All            429.    185.  
#> Matched (ESS)   87.4    93.93
#> Matched        185.    185.  
#> Unmatched      244.      0.  
#> Discarded        0.      0.  
#>